requirements.txt
gradio
rank_bm25
requests
unidecode
rapidfuzz
PyYAML
